Refactor | #22 | @lcomment | Entity 관련 수정
