+++ cakk-domain/src/main/java/com/cakk/domain/entity/shop/CakeShopOperation.java
@@ -1,4 +1,4 @@
-package com.cakk.domain.entity.cake;
+package com.cakk.domain.entity.shop;
 
 import java.time.LocalTime;
