+++ cakk-domain/src/main/java/com/cakk/domain/repository/jpa/CakeShopOperationJpaRepository.java
@@ -2,6 +2,6 @@ package com.cakk.domain.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cakk.domain.entity.cake.CakeShopOperation;
+import com.cakk.domain.entity.shop.CakeShopOperation;
 
 public interface CakeShopOperationJpaRepository extends JpaRepository<CakeShopOperation, Long> { }
