+++ cakk-domain/src/main/java/com/cakk/domain/repository/writer/CakeShopWriter.java
@@ -5,7 +5,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 
 import com.cakk.domain.annotation.Writer;
-import com.cakk.domain.entity.cake.CakeShopOperation;
+import com.cakk.domain.entity.shop.CakeShopOperation;
 import com.cakk.domain.entity.shop.CakeShop;
 import com.cakk.domain.entity.user.BusinessInformation;
 import com.cakk.domain.repository.jpa.BusinessInformationJpaRepository;
